Small optimization on the message processing routine.
The routine processing the incoming messages used to copy these messages
to a temporary buffer in order to block for the smallest amount of time
possible (indeed, processing the messages include the time to handle the
message by the receiver, time that can be quite long, depending on the
receiver).

Now instead of copying in a temporary buffer, we swap the buffers,
removing the costly copy operation.